pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spleen-seg"
version = "0.1.0"
description = "3D spleen segmentation with MONAI and PyTorch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "monai",
    "nibabel",
    "numpy",
    "onnxruntime-gpu",
    "torch",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
